Move ConnectionPool redirect tests to PoolManager.
