+++ test/with_dummyserver/test_poolmanager.py
@@ -1,5 +1,7 @@
 import unittest
+import io
 import json
+import time
 
 from nose.plugins.skip import SkipTest
 from dummyserver.server import HAS_IPV6
@@ -7,8 +9,10 @@ from dummyserver.testcase import (HTTPDummyServerTestCase,
                                   IPv6HTTPDummyServerTestCase)
 from urllib3.base import DEFAULT_PORTS
 from urllib3.poolmanager import PoolManager
-from urllib3.exceptions import MaxRetryError
-from urllib3.util.retry import Retry
+from urllib3.exceptions import (
+    MaxRetryError, NewConnectionError, UnrewindableBodyError
+)
+from urllib3.util.retry import Retry, RequestHistory
 
 
 class TestPoolManager(HTTPDummyServerTestCase):
@@ -213,6 +217,317 @@ class TestPoolManager(HTTPDummyServerTestCase):
 
         r = http.request('GET', 'http://%s:%s/' % (self.host, self.port))
         self.assertEqual(r.status, 200)
+
+    def test_cleanup_on_connection_error(self):
+        '''
+        Test that connections are recycled to the pool on
+        connection errors where no http response is received.
+        '''
+        poolsize = 3
+
+        with PoolManager(maxsize=poolsize, block=True) as http:
+            pool = http.connection_from_host(self.host, self.port)
+            self.assertEqual(pool.pool.qsize(), poolsize)
+
+            # force a connection error by supplying a non-existent
+            # url. We won't get a response for this  and so the
+            # conn won't be implicitly returned to the pool.
+            url = "%s/redirect" % self.base_url
+            self.assertRaises(MaxRetryError,
+                              http.request,
+                              'GET', url,
+                              fields={'target': '/'}, retries=0)
+
+            r = http.request('GET', url,
+                             fields={'target': '/'},
+                             retries=1)
+            r.release_conn()
+
+            # the pool should still contain poolsize elements
+            self.assertEqual(pool.pool.qsize(), poolsize)
+
+
+class TestRetry(HTTPDummyServerTestCase):
+
+    def setUp(self):
+        self.base_url = 'http://%s:%d' % (self.host, self.port)
+        self.base_url_alt = 'http://%s:%d' % (self.host_alt, self.port)
+
+    def test_max_retry(self):
+        with PoolManager() as http:
+            try:
+                r = http.request('GET', '%s/redirect' % self.base_url,
+                                 fields={'target': '/'},
+                                 retries=0)
+                self.fail("Failed to raise MaxRetryError exception, returned %r" % r.status)
+            except MaxRetryError:
+                pass
+
+    def test_disabled_retry(self):
+        """ Disabled retries should disable redirect handling. """
+        with PoolManager() as http:
+            r = http.request('GET', '%s/redirect' % self.base_url,
+                             fields={'target': '/'},
+                             retries=False)
+            self.assertEqual(r.status, 303)
+
+            r = http.request('GET', '%s/redirect' % self.base_url,
+                             fields={'target': '/'},
+                             retries=Retry(redirect=False))
+            self.assertEqual(r.status, 303)
+
+            self.assertRaises(NewConnectionError, http.request,
+                              'GET', 'http://thishostdoesnotexist.invalid/',
+                              timeout=0.001, retries=False)
+
+    def test_read_retries(self):
+        """ Should retry for status codes in the whitelist """
+        retry = Retry(read=1, status_forcelist=[418])
+
+        with PoolManager() as http:
+            resp = http.request('GET', '%s/successful_retry' % self.base_url,
+                                headers={'test-name': 'test_read_retries'},
+                                retries=retry)
+            self.assertEqual(resp.status, 200)
+
+    def test_read_total_retries(self):
+        """ HTTP response w/ status code in the whitelist should be retried """
+        headers = {'test-name': 'test_read_total_retries'}
+        retry = Retry(total=1, status_forcelist=[418])
+
+        with PoolManager() as http:
+            resp = http.request('GET', '%s/successful_retry' % self.base_url,
+                                headers=headers, retries=retry)
+            self.assertEqual(resp.status, 200)
+
+    def test_retries_wrong_whitelist(self):
+        """HTTP response w/ status code not in whitelist shouldn't be retried"""
+        retry = Retry(total=1, status_forcelist=[202])
+
+        with PoolManager() as http:
+            resp = http.request('GET', '%s/successful_retry' % self.base_url,
+                                headers={'test-name': 'test_wrong_whitelist'},
+                                retries=retry)
+            self.assertEqual(resp.status, 418)
+
+    def test_default_method_whitelist_retried(self):
+        """ urllib3 should retry methods in the default method whitelist """
+        retry = Retry(total=1, status_forcelist=[418])
+
+        with PoolManager() as http:
+            resp = http.request('OPTIONS', '%s/successful_retry' % self.base_url,
+                                headers={'test-name': 'test_default_whitelist'},
+                                retries=retry)
+            self.assertEqual(resp.status, 200)
+
+    def test_retries_wrong_method_list(self):
+        """Method not in our whitelist should not be retried, even if code matches"""
+        headers = {'test-name': 'test_wrong_method_whitelist'}
+        retry = Retry(total=1, status_forcelist=[418],
+                      method_whitelist=['POST'])
+
+        with PoolManager() as http:
+            resp = http.request('GET', '%s/successful_retry' % self.base_url,
+                                headers=headers, retries=retry)
+            self.assertEqual(resp.status, 418)
+
+    def test_read_retries_unsuccessful(self):
+        headers = {'test-name': 'test_read_retries_unsuccessful'}
+
+        with PoolManager() as http:
+            resp = http.request('GET', '%s/successful_retry' % self.base_url,
+                                headers=headers, retries=1)
+            self.assertEqual(resp.status, 418)
+
+    def test_retry_reuse_safe(self):
+        """ It should be possible to reuse a Retry object across requests """
+        headers = {'test-name': 'test_retry_safe'}
+        retry = Retry(total=1, status_forcelist=[418])
+
+        with PoolManager() as http:
+            resp = http.request('GET', '%s/successful_retry' % self.base_url,
+                                headers=headers, retries=retry)
+            self.assertEqual(resp.status, 200)
+            resp = http.request('GET', '%s/successful_retry' % self.base_url,
+                                headers=headers, retries=retry)
+            self.assertEqual(resp.status, 200)
+
+    def test_retry_return_in_response(self):
+        headers = {'test-name': 'test_retry_return_in_response'}
+        retry = Retry(total=2, status_forcelist=[418])
+
+        with PoolManager() as http:
+            resp = http.request('GET', '%s/successful_retry' % self.base_url,
+                                headers=headers, retries=retry)
+            self.assertEqual(resp.status, 200)
+            self.assertEqual(resp.retries.total, 1)
+            self.assertEqual(resp.retries.history,
+                             (RequestHistory('GET', '/successful_retry', None, 418, None),))
+
+    def test_retry_redirect_history(self):
+        with PoolManager() as http:
+            resp = http.request('GET', '%s/redirect' % self.base_url, fields={'target': '/'})
+            self.assertEqual(resp.status, 200)
+            self.assertEqual(resp.retries.history,
+                             (RequestHistory('GET',
+                                             self.base_url + '/redirect?target=%2F',
+                                             None,
+                                             303,
+                                             '/'),))
+
+    def test_multi_redirect_history(self):
+        with PoolManager() as http:
+            r = http.request('GET', '%s/multi_redirect' % self.base_url,
+                             fields={'redirect_codes': '303,302,200'},
+                             redirect=False)
+            self.assertEqual(r.status, 303)
+            self.assertEqual(r.retries.history, tuple())
+
+            r = http.request('GET', '%s/multi_redirect' % self.base_url, retries=10,
+                             fields={'redirect_codes': '303,302,301,307,302,200'})
+            self.assertEqual(r.status, 200)
+            self.assertEqual(r.data, b'Done redirecting')
+
+            expected = [(303, '/multi_redirect?redirect_codes=302,301,307,302,200'),
+                        (302, '/multi_redirect?redirect_codes=301,307,302,200'),
+                        (301, '/multi_redirect?redirect_codes=307,302,200'),
+                        (307, '/multi_redirect?redirect_codes=302,200'),
+                        (302, '/multi_redirect?redirect_codes=200')]
+            actual = [(history.status, history.redirect_location) for history in r.retries.history]
+            self.assertEqual(actual, expected)
+
+    def test_redirect_put_file(self):
+        """PUT with file object should work with a redirection response"""
+        retry = Retry(total=3, status_forcelist=[418])
+        # httplib reads in 8k chunks; use a larger content length
+        content_length = 65535
+        data = b'A' * content_length
+        uploaded_file = io.BytesIO(data)
+        headers = {'test-name': 'test_redirect_put_file',
+                   'Content-Length': str(content_length)}
+        url = '%s/redirect?target=/echo&status=307' % self.base_url
+
+        with PoolManager() as http:
+            resp = http.urlopen('PUT', url,
+                                headers=headers,
+                                retries=retry,
+                                body=uploaded_file)
+            self.assertEqual(resp.status, 200)
+            self.assertEqual(resp.data, data)
+
+
+class TestRetryAfter(HTTPDummyServerTestCase):
+    def setUp(self):
+        self.base_url = 'http://%s:%d' % (self.host, self.port)
+        self.base_url_alt = 'http://%s:%d' % (self.host_alt, self.port)
+
+    def test_retry_after(self):
+        url = "%s/retry_after" % self.base_url
+        with PoolManager() as http:
+            # Request twice in a second to get a 429 response.
+            r = http.request('GET', url,
+                             fields={'status': '429 Too Many Requests'},
+                             retries=False)
+            r = http.request('GET', url,
+                             fields={'status': '429 Too Many Requests'},
+                             retries=False)
+            self.assertEqual(r.status, 429)
+
+            r = http.request('GET', url,
+                             fields={'status': '429 Too Many Requests'},
+                             retries=True)
+            self.assertEqual(r.status, 200)
+
+            # Request twice in a second to get a 503 response.
+            r = http.request('GET', url,
+                             fields={'status': '503 Service Unavailable'},
+                             retries=False)
+            r = http.request('GET', url,
+                             fields={'status': '503 Service Unavailable'},
+                             retries=False)
+            self.assertEqual(r.status, 503)
+
+            r = http.request('GET', url,
+                             fields={'status': '503 Service Unavailable'},
+                             retries=True)
+            self.assertEqual(r.status, 200)
+
+            # Ignore Retry-After header on status which is not defined in
+            # Retry.RETRY_AFTER_STATUS_CODES.
+            r = http.request('GET', url,
+                             fields={'status': "418 I'm a teapot"},
+                             retries=True)
+            self.assertEqual(r.status, 418)
+
+    def test_redirect_after(self):
+        with PoolManager() as http:
+            r = http.request('GET', '%s/redirect_after' % self.base_url, retries=False)
+            self.assertEqual(r.status, 303)
+
+            t = time.time()
+            r = http.request('GET', '%s/redirect_after' % self.base_url)
+            self.assertEqual(r.status, 200)
+            delta = time.time() - t
+            self.assertTrue(delta >= 1)
+
+            t = time.time()
+            timestamp = t + 2
+            r = http.request('GET', self.base_url + '/redirect_after?date=' + str(timestamp))
+            self.assertEqual(r.status, 200)
+            delta = time.time() - t
+            self.assertTrue(delta >= 1)
+
+            # Retry-After is past
+            t = time.time()
+            timestamp = t - 1
+            r = http.request('GET', self.base_url + '/redirect_after?date=' + str(timestamp))
+            delta = time.time() - t
+            self.assertEqual(r.status, 200)
+            self.assertTrue(delta < 1)
+
+
+class TestFileBodiesOnRetryOrRedirect(HTTPDummyServerTestCase):
+    def setUp(self):
+        self.base_url = 'http://%s:%d' % (self.host, self.port)
+        self.base_url_alt = 'http://%s:%d' % (self.host_alt, self.port)
+
+    def test_retries_put_filehandle(self):
+        """HTTP PUT retry with a file-like object should not timeout"""
+        retry = Retry(total=3, status_forcelist=[418])
+        # httplib reads in 8k chunks; use a larger content length
+        content_length = 65535
+        data = b'A' * content_length
+        uploaded_file = io.BytesIO(data)
+        headers = {'test-name': 'test_retries_put_filehandle',
+                   'Content-Length': str(content_length)}
+
+        with PoolManager() as http:
+            resp = http.urlopen('PUT', '%s/successful_retry' % self.base_url,
+                                headers=headers,
+                                retries=retry,
+                                body=uploaded_file,
+                                redirect=False)
+            self.assertEqual(resp.status, 200)
+
+    def test_redirect_with_failed_tell(self):
+        """Abort request if failed to get a position from tell()"""
+        class BadTellObject(io.BytesIO):
+
+            def tell(self):
+                raise IOError
+
+        body = BadTellObject(b'the data')
+        url = '%s/redirect?target=/successful_retry' % self.base_url
+        # httplib uses fileno if Content-Length isn't supplied,
+        # which is unsupported by BytesIO.
+        headers = {'Content-Length': '8'}
+
+        with PoolManager() as http:
+            try:
+                http.urlopen('PUT', url, headers=headers, body=body)
+                self.fail('PUT successful despite failed rewind.')
+            except UnrewindableBodyError as e:
+                self.assertTrue('Unable to record file position for' in str(e))
 
 
 class TestIPv6PoolManager(IPv6HTTPDummyServerTestCase):
